Interactive mode subshell
We now have a full cel environment to implement interactive-only functions.

This allows us to extend bomshell with new CEL methods that only exist
when running in the interactive console.

We have one first method: help() which overloads the bomshell runtime
with a simple function that prints a help file.

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>