New render engine
bomshell has a new render engine that generates printable representations
of elements and native protobom types to use in the output of processes.
It is designed to have more than one implementation. The first one
(render.TTY) generates information about the expression results.

Ultimately, the plan for the package is to have high quality rendering
og the protobom sbom elements to use them in the output of different types
of applications (ie json, text, HTML, dot, mermaid, etc).

Signed-off-by: Adolfo García Veytia (Puerco) <puerco@chainguard.dev>